// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TransactionType = {
  DEPOSIT: 'DEPOSIT',
  EXPENSE: 'EXPENSE',
  INVESTMENT: 'INVESTMENT'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const TransactionCategory = {
  HOUSING: 'HOUSING',
  TRANSPORTATION: 'TRANSPORTATION',
  FOOD: 'FOOD',
  ENTERTAINMENT: 'ENTERTAINMENT',
  HEALTH: 'HEALTH',
  UTILITY: 'UTILITY',
  SALARY: 'SALARY',
  EDUCATION: 'EDUCATION',
  OTHER: 'OTHER'
} as const

export type TransactionCategory = (typeof TransactionCategory)[keyof typeof TransactionCategory]


export const TransactionPaymentMethod = {
  CREDIT_CARD: 'CREDIT_CARD',
  DEBIT_CARD: 'DEBIT_CARD',
  BANK_TRANSFER: 'BANK_TRANSFER',
  BANK_SLIP: 'BANK_SLIP',
  CASH: 'CASH',
  PIX: 'PIX',
  OTHER: 'OTHER'
} as const

export type TransactionPaymentMethod = (typeof TransactionPaymentMethod)[keyof typeof TransactionPaymentMethod]